rename component to basket
