+++ components/Basket/Basket.module.css
@@ -8,7 +8,8 @@
   padding: 0.5em 1em;
   background-color: #39f;
   color: white;
-  display: inline-block;
+  align-items: center;
+  display: inline-flex;
 }
 
 .content {
